chore($SQL): modify MySQL table structures
Modified Structures:

 - Add missing field: `exercise_id`;
 - Correct `parent_id` data type.

BREAKING CHANGE: update MySQL table structure (10:04 AM, Feb.11, 2020)

+++ src/main/java/com/jmsoftware/exrxnetcrawlerserver/kinesiologyglossary/domain/KinesiologyGlossaryPo.java
@@ -27,5 +27,5 @@ public class KinesiologyGlossaryPo {
     /**
      * The parent ID of kinesiology glossary.
      */
-    private Integer parentId;
+    private Long parentId;
 }
